pyproject.toml
[project]
name = "lamjav"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateus Kasuya",email = "mateusvbkasuya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "balldontlie (>=0.1.6,<0.2.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-cloud-storage (>=2.4,<3.2)",
    "google-cloud-bigquery (>=3.0.0,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "dbt-bigquery (>=1.10.0,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "camelot-py (>=1.0.0,<2.0.0)",
    "thefuzz (>=0.20.0,<1.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.12.2"
sqlfluff = "^3.4.2"
sqlfluff-templater-dbt = "^3.4.2"



[tool.taskipy.tasks]
fix = "ruff format . && ruff check . --fix && sqlfluff format . && sqlfluff fix ."
run = "poetry run python"
